+++ packages/components/src/button/types.ts
@@ -43,9 +43,15 @@ type BaseButtonProps = {
 	 */
 	children?: ReactNode;
 	/**
-	 * An accessible description for the button.
+	 * A visually hidden accessible description for the button.
 	 */
 	describedBy?: string;
+	/**
+	 * Whether to show the accessible description.
+	 *
+	 * @default false
+	 */
+	showDescription?: boolean;
 	/**
 	 * If provided, renders an Icon component inside the button.
 	 */
@@ -55,7 +61,7 @@ type BaseButtonProps = {
 	 *
 	 * @default 'left'
 	 */
-	iconPosition?: 'left' | 'right';
+	iconPosition?: 'left' | 'right' | 'top';
 	/**
 	 * If provided with `icon`, sets the icon size.
 	 * Please refer to the Icon component for more details regarding